feat: add extra goos support
syscall.Errno is not available is some GOOS causing compile errors
Move syscall specific code to separate file behind build tags making it
future proof and adding support for all goos supported by the go runtime
