+++ error_unix.go
@@ -15,8 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//go:build !windows
-// +build !windows
+//go:build unix
 
 package apm // import "go.elastic.co/apm/v2"
