+++ error_windows.go
@@ -15,6 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+//go:build windows
+
 package apm // import "go.elastic.co/apm/v2"
 
 import (
